lab 4. done

+++ src/main/java/ru/nstu/javafx_labs_lipatov_v2/data/IBehaviour.java
@@ -1,9 +1,9 @@
 package ru.nstu.javafx_labs_lipatov_v2.data;
 
 public interface IBehaviour {
-    void setX(int x);
-    void setY(int y);
-    int getX();
-    int getY();
+    void setX(double x);
+    void setY(double y);
+    double getX();
+    double getY();
     void move();
 }
